added support for decompressing tar.gz in pure go
