+++ unpack/unpack_zip.go
@@ -1,0 +1,54 @@
+package unpack
+
+import (
+	"archive/zip"
+	"fmt"
+	"io"
+	"os"
+	"path/filepath"
+	"strings"
+)
+
+func Unzip(src string, dir string) error {
+	archive, err := zip.OpenReader(src)
+	if err != nil {
+		return fmt.Errorf("unable to unzip: %v", err)
+	}
+	defer archive.Close()
+
+	for _, f := range archive.File {
+		fileName := f.Name
+
+		filePath := filepath.Join(dir, fileName)
+
+		if !strings.HasPrefix(filePath, filepath.Clean(dir)+string(os.PathSeparator)) {
+			return fmt.Errorf("invalid file path")
+		}
+		if f.FileInfo().IsDir() {
+			os.MkdirAll(filePath, os.ModePerm)
+			continue
+		}
+
+		if err := os.MkdirAll(filepath.Dir(filePath), os.ModePerm); err != nil {
+			return err
+		}
+
+		dstFile, err := os.OpenFile(filePath, os.O_WRONLY|os.O_CREATE|os.O_TRUNC, f.Mode())
+		if err != nil {
+			return err
+		}
+
+		fileInArchive, err := f.Open()
+		if err != nil {
+			return err
+		}
+
+		if _, err := io.Copy(dstFile, fileInArchive); err != nil {
+			return err
+		}
+
+		dstFile.Close()
+		fileInArchive.Close()
+	}
+	return nil
+}
